requirements.txt
boto3
psycopg[binary]
pyyaml
requests
python-dateutil
uvicorn
fastapi
lxml
jsonschema>=4.20
